+++ Runtime/Core/Platform/Platform.cs
@@ -1,4 +1,6 @@
 using System;
+using System.Collections.Generic;
+using System.Runtime.InteropServices;
 using UnityEngine;
 
 namespace Unity.AppUI.Core
@@ -43,11 +45,192 @@ namespace Unity.AppUI.Core
     }
 
     /// <summary>
+    /// A touch event received from a magic trackpad.
+    /// </summary>
+    /// <remarks>
+    /// Theses Touch events can be received from a magic trackpad on macOS.
+    /// </remarks>
+    public struct TrackPadTouch
+    {
+        /// <summary>
+        /// The unique identifier of the touch.
+        /// </summary>
+        public int fingerId { get; }
+
+        /// <summary>
+        /// The position of the touch in normalized coordinates.
+        /// </summary>
+        public Vector2 position { get; }
+
+        /// <summary>
+        /// The number of taps. This is always 1 for a trackpad.
+        /// </summary>
+        public int tapCount { get; }
+
+        /// <summary>
+        /// The delta position of the touch since the last frame.
+        /// </summary>
+        public Vector2 deltaPos { get; }
+
+        /// <summary>
+        /// The delta time since the last frame.
+        /// </summary>
+        public float deltaTime { get; }
+
+        /// <summary>
+        /// The phase of the touch.
+        /// </summary>
+        public TouchPhase phase { get; }
+
+        /// <summary>
+        /// Constructor.
+        /// </summary>
+        /// <param name="fingerId"> The unique identifier of the touch.</param>
+        /// <param name="position"> The position of the touch in normalized coordinates.</param>
+        /// <param name="tapCount"> The number of taps. This is always 1 for a trackpad.</param>
+        /// <param name="deltaPos"> The delta position of the touch since the last frame.</param>
+        /// <param name="deltaTime"> The delta time since the last frame.</param>
+        /// <param name="phase"> The phase of the touch.</param>
+        public TrackPadTouch(int fingerId, Vector2 position, int tapCount, Vector2 deltaPos, float deltaTime,
+            TouchPhase phase)
+        {
+            this.fingerId = fingerId;
+            this.position = position;
+            this.tapCount = tapCount;
+            this.deltaPos = deltaPos;
+            this.deltaTime = deltaTime;
+            this.phase = phase;
+        }
+    }
+
+    /// <summary>
+    /// A magnification gesture received from a magic trackpad.
+    /// </summary>
+    public struct MagnificationGesture : IEquatable<MagnificationGesture>
+    {
+        /// <summary>
+        /// The magnification delta of the gesture since the last frame.
+        /// </summary>
+        public float deltaMagnification { get; }
+
+        /// <summary>
+        /// The phase of the gesture.
+        /// </summary>
+        public TouchPhase phase { get; }
+
+        /// <summary>
+        /// Constructor.
+        /// </summary>
+        /// <param name="deltaMagnification">The magnification delta of the gesture since the last frame.</param>
+        /// <param name="phase">The phase of the gesture.</param>
+        public MagnificationGesture(float deltaMagnification, TouchPhase phase)
+        {
+            this.phase = phase;
+            this.deltaMagnification = deltaMagnification;
+        }
+
+        public bool Equals(MagnificationGesture other)
+        {
+            return Mathf.Approximately(deltaMagnification, other.deltaMagnification) && phase == other.phase;
+        }
+
+        public override bool Equals(object obj)
+        {
+            return obj is MagnificationGesture other && Equals(other);
+        }
+
+        public override int GetHashCode()
+        {
+            return HashCode.Combine(deltaMagnification, (int) phase);
+        }
+        
+        public static bool operator ==(MagnificationGesture left, MagnificationGesture right)
+        {
+            return left.Equals(right);
+        }
+        
+        public static bool operator !=(MagnificationGesture left, MagnificationGesture right)
+        {
+            return !left.Equals(right);
+        }
+    }
+
+    /// <summary>
+    /// A pan gesture received from a magic trackpad.
+    /// </summary>
+    public struct PanGesture : IEquatable<PanGesture>
+    {
+        /// <summary>
+        /// The delta position of the touch since the last frame.
+        /// </summary>
+        public Vector2 deltaPos { get; }
+
+        /// <summary>
+        /// The phase of the gesture.
+        /// </summary>
+        public TouchPhase phase { get; }
+        
+        /// <summary>
+        /// The position of the touch in normalized coordinates.
+        /// </summary>
+        public Vector2 position { get; }
+        
+        /// <summary>
+        /// Constructor.
+        /// </summary>
+        /// <param name="position">The position of the touch in normalized coordinates.</param>
+        /// <param name="deltaPos">The delta position of the touch since the last frame.</param>
+        /// <param name="phase">The phase of the gesture.</param>
+        public PanGesture(Vector2 position, Vector2 deltaPos, TouchPhase phase)
+        {
+            this.position = position;
+            this.deltaPos = deltaPos;
+            this.phase = phase;
+        }
+
+        public bool Equals(PanGesture other)
+        {
+            return deltaPos.Equals(other.deltaPos) && phase == other.phase && position.Equals(other.position);
+        }
+
+        public override bool Equals(object obj)
+        {
+            return obj is PanGesture other && Equals(other);
+        }
+
+        public override int GetHashCode()
+        {
+            return HashCode.Combine(deltaPos, position, (int) phase);
+        }
+        
+        public static bool operator ==(PanGesture left, PanGesture right)
+        {
+            return left.Equals(right);
+        }
+        
+        public static bool operator !=(PanGesture left, PanGesture right)
+        {
+            return !left.Equals(right);
+        }
+    }
+    
+    [StructLayout(LayoutKind.Sequential)]
+    struct PlatformTouchEvent 
+    {
+        public byte touchId;
+        public byte phase;
+        public float normalizedX;
+        public float normalizedY;
+        public float deviceWidth;
+        public float deviceHeight;
+    }
+
+    /// <summary>
     /// Utility methods and properties related to the Target Platform.
     /// </summary>
     public static partial class Platform
     {
-        const float k_BaseDpi = 96f;
+        public const float baseDpi = 96f;
 
         /// <summary>
         /// The DPI value that should be used in UI-Toolkit PanelSettings
@@ -63,7 +246,7 @@ namespace Unity.AppUI.Core
 #if UNITY_EDITOR_WIN || UNITY_STANDALONE_WIN
             // On Windows we can use a value of 96dpi because UI Toolkit scales correctly the UI based on
             // Operating System's DPI and ScaleFactor changes.
-            return k_BaseDpi;
+            return baseDpi;
 #else
                 return Screen.dpi / mainScreenScale;
 #endif
@@ -93,12 +276,12 @@ namespace Unity.AppUI.Core
                 // Scale factor with a 4k monitor is 150%, which gives 96 * 1.5 = 144dpi.
                 // Unity Engine sets the DPI awareness per monitor, so the UI will scale automatically :
                 // https://docs.microsoft.com/en-us/windows/win32/api/windef/ne-windef-dpi_awareness
-                return Screen.dpi / k_BaseDpi;
+                return Screen.dpi / baseDpi;
 #endif
             }
         }
 
-        static event Action<string> s_SystemThemeChangedInternal;
+        static event Action<string> systemThemeChangedInternal;
 
         /// <summary>
         /// Event triggered when the system theme changes.
@@ -107,14 +290,14 @@ namespace Unity.AppUI.Core
         {
             add
             {
-                s_SystemThemeChangedInternal += value;
+                systemThemeChangedInternal += value;
                 if (value != null)
                     EnableThemePolling();
             }
             remove
             {
-                s_SystemThemeChangedInternal -= value;
-                if (s_SystemThemeChangedInternal == null)
+                systemThemeChangedInternal -= value;
+                if (systemThemeChangedInternal == null)
                     DisableThemePolling();
             }
         }
@@ -152,10 +335,262 @@ namespace Unity.AppUI.Core
             var currentTheme = systemTheme;
             if (currentTheme != s_PreviousSystemTheme)
             {
-                s_SystemThemeChangedInternal?.Invoke(currentTheme);
+                systemThemeChangedInternal?.Invoke(currentTheme);
                 s_PreviousSystemTheme = currentTheme;
             }
         }
+
+        static TrackPadTouch s_Touch0;
+        static TrackPadTouch s_Touch1;
+
+        static bool s_TwoFingersUsed;
+
+        static Vector2 s_StartPos0;
+        static Vector2 s_StartPos1;
+
+        static Vector2 s_DeltaPos0;
+        static Vector2 s_DeltaPos1;
+
+        enum GestureType 
+        {
+            Unknown,
+            Pan,
+            Mag
+        }
+
+        static GestureType s_Gesture;
+
+        static float s_LastMag;
+
+        static Vector2 s_LastPanPos;
+
+        static float s_StartDistance;
+
+        const float k_RecognitionThreshold = 0.065f;
+
+        const float k_PanRecognitionDotThreshold = 0.98f;
+
+        /// <summary>
+        /// Polls the gestures and triggers gesture events if a gesture has been received.
+        /// </summary>
+        internal static void PollGestures()
+        {
+            panGestureChangedThisFrame = false;
+            magnificationGestureChangedThisFrame = false;
+            
+            var touches = GetTrackpadTouches();
+
+            foreach (var touch in touches)
+            {
+                if (touch.phase == TouchPhase.Began)
+                {
+                    // check if we already have another touch
+                    if (s_Touch0.fingerId == -1 || s_Touch0.fingerId == touch.fingerId)
+                    {
+                        s_Touch0 = touch;
+                    }
+                    else if (s_Touch1.fingerId == -1 || s_Touch1.fingerId == touch.fingerId)
+                    {
+                        if (touch.fingerId == s_Touch0.fingerId)
+                            Debug.LogError("The second finger ID can't be the same as the first one");
+                        s_Touch1 = touch;
+                        s_TwoFingersUsed = true;
+                        s_StartPos0 = s_Touch0.position;
+                        s_StartPos1 = s_Touch1.position;
+                        s_DeltaPos0 = Vector2.zero;
+                        s_DeltaPos1 = Vector2.zero;
+                    }
+                    
+                    if (s_Touch0.fingerId != touch.fingerId && s_Touch1.fingerId != touch.fingerId && s_TwoFingersUsed)
+                    {
+                        // we can deactivate here since more than 2 fingers are used
+                        AbortGesture();
+                    }
+                }
+                else if (touch.phase is TouchPhase.Canceled or TouchPhase.Ended)
+                {
+                    // check if the touch is one of the tracked ones
+                    if (touch.fingerId == s_Touch0.fingerId || touch.fingerId == s_Touch1.fingerId)
+                        AbortGesture(touch.phase);
+                }
+
+                if (s_TwoFingersUsed && touch.phase is TouchPhase.Moved or TouchPhase.Stationary)
+                {
+                    // store touch data if its one of the tracked ones
+                    if (touch.fingerId == s_Touch0.fingerId)
+                    {
+                        s_Touch0 = touch;
+                    }
+                    else if (touch.fingerId == s_Touch1.fingerId)
+                    {
+                        s_Touch1 = touch;
+                    }
+                }
+            }
+
+            if (s_TwoFingersUsed)
+            {
+                // compute deltas
+                s_DeltaPos0 = s_Touch0.position - s_StartPos0;
+                s_DeltaPos1 = s_Touch1.position - s_StartPos1;
+
+                if (s_Gesture == GestureType.Unknown)
+                {
+                    // need to recognize the movement
+                    if (s_DeltaPos0.magnitude > k_RecognitionThreshold && s_DeltaPos1.magnitude > k_RecognitionThreshold)
+                    {
+                        s_LastMag = 1f;
+                        s_StartDistance = Vector2.Distance(s_Touch0.position, s_Touch1.position);
+                        s_LastPanPos = s_Touch0.position;
+
+                        // can recognize
+                        if (Vector2.Dot(s_DeltaPos0.normalized, s_DeltaPos1.normalized) > k_PanRecognitionDotThreshold)
+                        {
+                            // pan
+                            s_Gesture = GestureType.Pan;
+                            panGesture = new PanGesture(s_Touch0.position, Vector2.zero, TouchPhase.Began);
+                        }
+                        else 
+                        {
+                            // mag
+                            s_Gesture = GestureType.Mag;
+                            magnificationGesture = new MagnificationGesture(0f, TouchPhase.Began);
+                        }
+                    }
+                }
+                else if (s_Gesture == GestureType.Mag)
+                {
+                    var d = Vector2.Distance(s_Touch0.position, s_Touch1.position);
+                    var mag = d / s_StartDistance;
+                    var magDelta = mag - s_LastMag;
+                    s_LastMag = mag;
+                    magnificationGesture = new MagnificationGesture(magDelta, TouchPhase.Moved);
+                }
+                else if (s_Gesture == GestureType.Pan)
+                {
+                    var pos = s_Touch0.position;
+                    var panDelta = pos - s_LastPanPos;
+                    s_LastPanPos = pos;
+                    panGesture = new PanGesture(pos, panDelta, TouchPhase.Moved);
+                }
+            }
+        }
+
+        static float s_LastTime;
+        static int s_LastFrame;
+
+        static readonly Dictionary<int, TrackPadTouch> k_PrevTouches = new Dictionary<int, TrackPadTouch>();
+        static readonly List<TrackPadTouch> k_FrameTouches = new List<TrackPadTouch>();
+
+        static List<TrackPadTouch> GetTrackpadTouches() 
+        {
+            s_LastTime = Time.unscaledTime;
+            if (s_LastFrame != Time.frameCount)
+            {
+                s_LastFrame = Time.frameCount;
+                
+                k_PrevTouches.Clear();
+                foreach (var touch in k_FrameTouches)
+                {
+                    k_PrevTouches[touch.fingerId] = touch;
+                }
+
+                k_FrameTouches.Clear();
+                
+                PlatformTouchEvent e;
+                e.touchId = 0;
+                e.phase = 0;
+                e.normalizedX = 0;
+                e.normalizedY = 0;
+                e.deviceWidth = 0;
+                e.deviceHeight = 0;
+                
+                while (ReadTouch(ref e))
+                {
+                    var screenPos = new Vector2(e.normalizedX, e.normalizedY);
+                    var deltaPos = new Vector2(0, 0);
+
+                    if (k_PrevTouches.TryGetValue(e.touchId, out var prevTouch))
+                        deltaPos = screenPos - prevTouch.position;
+
+                    var timeDelta = Time.unscaledTime - s_LastTime;
+                    var phase = e.phase switch 
+                    {
+                        0 => TouchPhase.Began,
+                        1 => TouchPhase.Moved,
+                        2 => TouchPhase.Ended,
+                        3 => TouchPhase.Canceled,
+                        4 => TouchPhase.Stationary,
+                        _ => TouchPhase.Ended
+                    };
+                    var newTouch = new TrackPadTouch(e.touchId, screenPos, 1, deltaPos, timeDelta, phase);
+                    k_FrameTouches.Add(newTouch);
+                }
+
+                s_LastTime = Time.unscaledTime;
+            }
+            
+            return k_FrameTouches;
+        }
+
+        static bool ReadTouch(ref PlatformTouchEvent touch)
+        {
+#if UNITY_EDITOR_OSX || UNITY_STANDALONE_OSX
+            return ReadTouchEvent(ref touch);
+#else 
+            return false;
+#endif
+        }
+
+        static void AbortGesture(TouchPhase phase = TouchPhase.Canceled)
+        {
+            s_TwoFingersUsed = false;
+            s_Gesture = GestureType.Unknown;
+            s_Touch0 = new TrackPadTouch(-1, Vector2.zero, 0, Vector2.zero, 0, TouchPhase.Canceled);
+            s_Touch1 = new TrackPadTouch(-1, Vector2.zero, 0, Vector2.zero, 0, TouchPhase.Canceled);
+            panGesture = new PanGesture(Vector2.zero, Vector2.zero, phase);
+            magnificationGesture = new MagnificationGesture(0f, phase);
+        }
+        
+        public static event Action<PanGesture> panGestureChanged;
+        
+        public static event Action<MagnificationGesture> magnificationGestureChanged;
+        
+        public static bool panGestureChangedThisFrame { get; set; }
+        
+        public static bool magnificationGestureChangedThisFrame { get; set; }
+
+        public static PanGesture panGesture
+        {
+            get => s_PanGesture;
+            private set
+            {
+                if (s_PanGesture != value)
+                {
+                    s_PanGesture = value;
+                    panGestureChangedThisFrame = true;
+                    panGestureChanged?.Invoke(value);
+                }
+            }
+        }
+
+        public static MagnificationGesture magnificationGesture
+        {
+            get => s_MagnificationGesture;
+            private set
+            {
+                if (s_MagnificationGesture != value)
+                {
+                    s_MagnificationGesture = value;
+                    magnificationGestureChangedThisFrame = true;
+                    magnificationGestureChanged?.Invoke(value);
+                }
+            }
+        }
+
+        static PanGesture s_PanGesture;
+
+        static MagnificationGesture s_MagnificationGesture;
 
         /// <summary>
         /// The current system theme.
@@ -194,6 +629,19 @@ namespace Unity.AppUI.Core
             else
                 Debug.LogWarning("Haptic Feedbacks are not supported on the current platform.");
 #endif
+        }
+
+        static void OnDomainUnloaded(object sender, EventArgs e)
+        {
+#if UNITY_EDITOR_OSX || UNITY_STANDALONE_OSX
+            DestroyTrackingObject();
+#endif
+        }
+        
+        static Platform()
+        {
+            AppDomain.CurrentDomain.DomainUnload -= OnDomainUnloaded;
+            AppDomain.CurrentDomain.DomainUnload += OnDomainUnloaded;
         }
     }
 }