+++ Documentation~/whats-new.md
@@ -2,7 +2,7 @@
 uid: whats-new
 ---
 
-# What's New in **0.3.2**
+# What's New in **0.3.3**
 
 This section contains information about new features, improvements, and issues fixed.
 
@@ -10,6 +10,13 @@ For a complete list of changes made, refer to the **Changelog** page.
 
 The main updates in this release include:
 
+### Added 
+
+- Added Magic Trackpad gesture support for MacOS.
+- Added `PanGesture` and `MagnificationGesture` events for UITK dispatcher.
+
 ### Fixed
 
-- Fixed NavAction being added twice in NavGraph when deleting a linked NavDestination.
+- Fixed `Pressable` post-processing for disabled targets.
+- Fixed visual content rendering synchronization in `Progress` and `ExVisualElement` components.
+- Fixed Animation Recycling in `NavHost` component.